Fully enable shift-tabbing in Modal components (#170)
Previously shift-tab only worked until focus reached the first
focusable element in the modal. Hitting shift-tab when focus was on the
first element in the modal resulted in effectively a no-op as focus was
programmatically set to the first focusable element. With this change a
shift-tab while the first element on the modal has focus will move focus
to the _last_ element in the modal, as you would expect.