Add new rule to checking contact prop length
Since the autofixable rule is going to be re-writing code to use the
va-telephone component, we want to be able to notify developers when the
contact prop doesn't match the expectations.

This is a separate rule since the ESLint best practices [1] say to keep
the fixes as small as possible. Having one rule rewrite code to use a
`<va-telephone>` instead of an `<a>` _and_ also having it rewrite the
value of the contact prop by possibly removing a number seemed like a
bad idea.

- 1. https://eslint.org/docs/latest/developer-guide/working-with-rules#applying-fixes